[ArticleReplySummary] extract colored reply summary into component
